Fixed issue around EOI filtering using non signed agreements
The filtering was broken due to the changes to the get legal entities api call. Now we return both legal entities with signed and non signed agreements. The filter only needs to look at signed agreements so we have modified the logic to filter out legal entities with unsigned agreements.